requirements.txt
# Orchestration & AI
langgraph==0.0.26
google-genai==0.3.0
mcp[cli]==1.2.0

# External Tools
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
ddgs>=9.0.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
tenacity>=8.2.0

# The New API Layer
fastapi==0.110.0
uvicorn>=0.30.0
python-multipart==0.0.9